requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
bandit==1.7.9
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
confection==0.1.5
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.20.0
debugpy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/debugpy_1699267934478/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.8.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastjsonschema==2.20.0
fastparquet==2024.5.0
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719171233697/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1717717537056/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1717182742060/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_80l8bw91xp/croot/jupyter_core_1718818299834/work
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.4.0
langsmith==0.1.48
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numpy==1.26.4
openai==1.21.2
orjson==3.10.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy==0.5.6
pbr==6.0.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
plotly==5.24.0
preshed==3.0.9
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
psutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/psutil_1699246686105/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.18.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c7mhasdker/croot/pyzmq_1705605099593/work
referencing==0.35.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.25.0
setuptools==74.1.2
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
squarify==0.4.4
srsly==2.4.8
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.41.3
statsmodels==0.14.3
stevedore==5.2.0
tenacity==8.2.3
thinc==8.2.5
thinc_apple_ops==0.1.5
threadpoolctl==3.5.0
tiktoken==0.6.0
tornado @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_7e0ruwlfee/croot/tornado_1718740118099/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritools==4.0.3
urlextract==1.9.0
urllib3==2.2.1
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
weasel==0.4.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
